Add a demo website made with HTML and CSS
This website was created for the "Construção de Página Web" course, as
the first evaluation project. No Javascript was used on it.

index_tableful.html is the first exercise, it's the index page for the
website using a tableful layout system to arrange elements on the
screen. The links on it send the user to the other pages.
All other webpages use tableless layouts, with CSS for fine control over
the positioning of the page elements. The index page was re-written with
this method as well.